refactor(graphql-subscriptions) Extracted common implementation of WS and GraphQL Subscription messages panel
